Merge in SEQ/paws from feature/PTSD-1264-excessNoiseCv-float to develop

* commit 'bfe9abd547590dfe20ba7774a89c4a3ea4c9cd77':
  excessNoiseCv -> float

+++ pkg/web/models.go
@@ -201,7 +201,7 @@ type AnalogObject struct {
 
 	// Coefficient of variation of excess noise
 	// Example: 3
-	ExcessNoiseCv int32 `json:"excessNoiseCv"`
+	ExcessNoiseCv float64 `json:"excessNoiseCv"`
 
 	// Average time in seconds of the width of pulses of this analog
 	// Example: 0.11
